==> notebooks/01.1.geo.heatmap.ipynb <==
import pandas as pd
import os
import folium
import subprocess
from sklearn import preprocessing

print(folium.__version__)

df = pd.read_csv('data/campaign_contributions_for_map.tsv', sep='\t')
df.sort_values('Amount', ascending=False).head()

df.info()

valid_heatmap_df = df[["lat", "lon", "Amount"]]
valid_heatmap_df.head()

## Heatmap

# convert pandas.DataFrame to numpy.array
data = valid_heatmap_df[["lat", "lon", "Amount"]].values

# normalize weight values to [0, 1]
min_max_scaler = preprocessing.MinMaxScaler()
data[:, [2]] = min_max_scaler.fit_transform(data[:, [2]])

# map params
location = [41.0, -75.5]
zoom_start = 8
# heatmap params
# layer name
name = "layer01"
max_val = data[:, 2].max()
min_opacity = 0.9
blur = 20
radius = 12
help(folium.Map)

# create map
# m = folium.Map(location=location, zoom_start=zoom_start, )
m = folium.Map(location=location, tiles="stamentoner", zoom_start=zoom_start)
hm = HeatMap(data.tolist(), min_opacity=min_opacity, blur=blur, radius=radius)

hm.add_to(m)
# folium.GeoJson(district23).add_to(hmap)
# hmap.add_child(hm_wide)

m.save(os.path.join('results', 'Heatmap_1.html'))
hm.save(os.path.join('results', 'geo_heatmap.html'))


# Saving png's of Folium Maps

https://github.com/python-visualization/folium/issues/35

These aren't tested, just pasted here as a starting point should someone need them.

### Using cutycapt

# save to png file
outdir = "results" # this directory has to exist..
#hmap.save("tmp.html")  # heatmap saved above
url = "file://{}/results/geo_heatmap.html".format(os.getcwd())
outfn = os.path.join(outdir,"geo_heatmap.png")
subprocess.check_call(["cutycapt","--url={}".format(url), "--out={}".format(outfn)])

### Using selenium

import os
import time
from selenium import webdriver

fn = os.path.join('results', 'geo_heatmap.html')
tmpurl='file://{path}/{mapfile}'.format(path=os.getcwd(),mapfile=fn)
#hmap.save("tmp.html")  # heatmap saved above

browser = webdriver.Firefox()
browser.get(tmpurl)
time.sleep(5)  #Give the map tiles some time to load
browser.save_screenshot('geo_heatmap.png')
browser.quit()
