==> notebooks/04_Bar Plots.ipynb <==
import os, sys, inspect
import numpy as np
import matplotlib.pyplot as plt
# add parent dir to system dir
currdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
rootdir = os.path.dirname(currdir)
sys.path.insert(0, rootdir)

from src.plot import bar
import pandas as pd

# Importing file
data = "../data/int/realestate_demographics.csv"
# Reading the file
df_data = pd.read_csv(data)
df_data

# bar_plot_1 = df_data[["Zip Code", "Year", "Month", "Home Value Index", "Median Income (dollars)"]]
bar_plot_1 = df_data[["Zip Code", "Year", "Month", "Home Value Index"]]
bar_plot_1

bar_plot_g = bar_plot_1.groupby(["Zip Code", "Year"])
bar_plot_g.mean()


test_1 = bar_plot_1.loc[bar_plot_1["Zip Code"]==78748, :]
test_1

test_1_g = test_1.groupby(["Zip Code", "Year"])
test_1_g.mean()

aaa = round(test_1_g["Year"].max()/1, 0)
aaa

figsize = (20, 8)
fig = plt.figure(figsize=figsize)
ax = fig.subplots(1, 1)

labels = test_1_g["Year"].max()
height = test_1_g["Home Value Index"].mean()

title = "title"
ylabel = "ylabel"

bar(height, ticklabels=labels,
    ylabel=ylabel,
    ax=ax)

# figsize = (10, 8)
# fig = plt.figure(figsize=figsize)
# ax = fig.subplots(1, 1)

# labels = 'Frogs', 'Hogs', 'Dogs', 'Logs'
# height = [15, 30, 45, 10]

# title = "title"
# ylabel = "ylabel"

# bar(height, ticklabels=labels,
#     ylabel=ylabel,
#     ax=ax)
