==> notebooks/07.1.mortgage.rates.ipynb <==
import pandas as pd
import os, sys, inspect
import numpy as np
import matplotlib.pyplot as plt
# add parent dir to system dir
currdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
rootdir = os.path.dirname(currdir)
sys.path.insert(0, rootdir)

ffname = os.path.join(rootdir, "data", "raw", "FMAC-30us.csv")
mortgage_30_rates = pd.read_csv(ffname)
mortgage_30_rates['year'] = pd.DatetimeIndex(mortgage_30_rates['Date']).year
mortgage_30_rates['month'] = pd.DatetimeIndex(mortgage_30_rates['Date']).month
mortgage_30_rates = mortgage_30_rates.rename(index=str, columns={"Value": "Mortgage Rate"})
mortgage_30_rates = mortgage_30_rates.drop(labels="Date", axis=1)

output = os.path.join(rootdir, "data", "int", "30_year_rates.csv")
mortgage_30_rates.to_csv(output, index=False)
mortgage_30_rates.head()

ffname = os.path.join(rootdir, "data", "raw", "FMAC-15us.csv")
mortgage_15_rates = pd.read_csv(ffname)
mortgage_15_rates['year'] = pd.DatetimeIndex(mortgage_15_rates['Date']).year
mortgage_15_rates['month'] = pd.DatetimeIndex(mortgage_15_rates['Date']).month
mortgage_15_rates = mortgage_15_rates.rename(index=str, columns={"Value": "Mortgage Rate"})
mortgage_15_rates = mortgage_15_rates.drop(labels="Date", axis=1)

output = os.path.join(rootdir, "data", "int", "15_year_rates.csv")
mortgage_15_rates.to_csv(output, index=False)
mortgage_15_rates.head()
