==> notebooks/08.1.read.school.data.ipynb <==
import pandas as pd
import requests
import os, sys, inspect
import numpy as np
import usaddress
import matplotlib.pyplot as plt
# add parent dir to system dir
currdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
rootdir = os.path.dirname(currdir)
sys.path.insert(0, rootdir)

from src.gschools import GreatSchoolsApi

# init. GreatSchoolsApi instance
gsp = GreatSchoolsApi()

# get schools in Austin (return type dataframe)
state = "tx"
city = "austin"
school_type = None
school_level = None
schools_df = gsp.browse_schools(state, city, school_type, school_level, limit=1000)

schools_df["Zip Code"] = ''
for i in range(len(schools_df)):
    schools_df.iloc[i,23] = list(usaddress.tag(schools_df["address"][i])[0].values())[-1]

schools_df["parentRating"].fillna(0.00, inplace=True)
schools_df["gsRating"].fillna(0.00, inplace=True)
schools_df['parentRating'] = schools_df['parentRating'].apply(pd.to_numeric)
schools_df['gsRating'] = schools_df['gsRating'].apply(pd.to_numeric)

schools_grpzip = schools_df.groupby("Zip Code")
schools_byZip = schools_grpzip.mean().reset_index().round(2)

# # get school score (return type dict)
# state = "tx"
# gsid = schools_df["gsId"][0]
# school_score = gsp.school_test_score(state, gsid)

# # find nearby schools (return type datafarme)
# state = "tx"
# zip = 78703
# radius = 1

# nearbyschools_df = gsp.nearby_schools(state, zip=zip, radius=radius)

# save schools data to csv file
fname = f"great_scools_{city}.csv"
ffname = os.path.join(rootdir, "data", "int", fname)
schools_byZip.to_csv(ffname,index=False)

schools_df.shape


schools_df.columns
